Change Distance to double, add mapper tests
Round distance for the view model.
Fixed DeliveryYear - DeliveryYearViewModel.

+++ sfa.Tl.Marketing.Communication.Models/Dto/ProviderLocation.cs
@@ -10,7 +10,7 @@ namespace sfa.Tl.Marketing.Communication.Models.Dto
         public string Town { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
-        public int DistanceInMiles { get; set; }
+        public double DistanceInMiles { get; set; }
         public IEnumerable<Qualification> Qualification2020 { get; set; }
         public IEnumerable<Qualification> Qualification2021 { get; set; }
         public IEnumerable<DeliveryYear> Qualifications { get; set; }
